requirements.txt
pandas
numpy
ccxt
pyyaml
matplotlib
tqdm
